Make QA group green: add Pkg compat entry; fix module-in-if load error
The QA group failed two ways:

1. Aqua.test_deps_compat flagged the `Pkg` test extra (added for the
   grouped Core group) as lacking a [compat] entry. Add `Pkg = "1"` to
   the root Project.toml [compat] — the genuine fix Aqua requests, not a
   silenced check.

2. With that resolved, runtests.jl failed to lower with
   `syntax: "module" expression not at top level`. The grouping refactor
   wrapped the Core test body (which declares `module GlobalsTest` and
   `module NotInitTest`) in `if GROUP == "All" || GROUP == "Core" ... end`;
   `module` is illegal inside a non-top-level `if`, so the whole file
   failed to load for every GROUP (including QA and All). Move the Core
   body verbatim into test/core_tests.jl and `include` it from the `if`
   branch, where the included file's contents are top level again. No
   Core test logic changed.

Verified locally on julia 1.11: GROUP=QA passes (Aqua 10/10) and
GROUP=Core passes.
