+++ Project.toml
@@ -12,6 +12,7 @@ Serialization = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
 Aqua = "0.8"
 BenchmarkTools = "1"
 ExprTools = "0.1"
+Pkg = "1"
 SHA = "0.7, 1"
 Serialization = "1.10"
 Test = "1"
